revert change for storagekeyname
